proj3

+++ src/vm/page.h
@@ -6,6 +6,8 @@
 #include "filesys/file.h"
 #include "filesys/off_t.h"
 #include "vm/frame.h"
+#include "threads/vaddr.h"
+#include <stdio.h>
 
 struct s_page{
     void *kpage;
@@ -20,7 +22,7 @@ struct s_page{
     // informations from process.c load_segment function
 
 
-    int swap_index;
+    size_t swap_index;
     int page_id;
     int fid;
 }
